Переносит получение информации о пользователе в PlatformUser
