+++ project/apps/users/src/app/authentication/authentication.http
@@ -3,8 +3,8 @@ POST http://localhost:3000/api/auth/register HTTP/1.1
 Content-Type: application/json
 
 {
-  "name": "Alice",
-  "email": "@gmail.com",
+  "name": "John",
+  "email": "john@gmail.com",
   "city": "Москва",
   "password": "123456",
   "role": "Исполнитель",
@@ -13,17 +13,12 @@ Content-Type: application/json
 
 ###
 
-# Get user by ID
-GET http://localhost:3000/api/auth/642e888aebee2196eac2ae22 HTTP/1.1
-
-###
-
 # Login user by email and password
 POST http://localhost:3000/api/auth/login HTTP/1.1
 Content-Type: application/json
 
 {
-  "email": "alice@gmail.com",
+  "email": "john@gmail.com",
   "password": "123456"
 }
