+++ project/apps/users/src/app/platform-user/platform-user.service.ts
@@ -1,0 +1,13 @@
+import { Injectable } from '@nestjs/common';
+import { PlatformUserRepository } from '../platform-user/platform-user.repository';
+
+@Injectable()
+export class PlatformUserService {
+  constructor(
+    private readonly platformUserRepository: PlatformUserRepository
+  ) {}
+
+  public async getUser(id: string) {
+    return this.platformUserRepository.findById(id);
+  }
+}
